Improved: Moved around the namespaces for the new functionality
