+++ projects/api/dxt-lossless-transform-bc1-api/src/c_api/determine_optimal_transform/mod.rs
@@ -2,10 +2,8 @@
 
 pub mod builder;
 
-pub use builder::*;
-
 use crate::c_api::error::{Dltbc1ErrorCode, Dltbc1Result};
-use crate::c_api::transform_context::{Dltbc1TransformContext, get_context_mut};
+use crate::c_api::transform::transform_context::{Dltbc1TransformContext, get_context_mut};
 use crate::determine_optimal_transform::determine_optimal_transform;
 use crate::{Bc1Error, transform::Bc1TransformOptionsBuilder};
 use core::slice;
