+++ projects/api/dxt-lossless-transform-bc1-api/src/c_api/mod.rs
@@ -6,10 +6,3 @@
 pub mod determine_optimal_transform;
 pub mod error;
 pub mod transform;
-pub mod transform_context;
-
-// Re-export all C API functions
-pub use determine_optimal_transform::*;
-pub use error::*;
-pub use transform::*;
-pub use transform_context::*;
