+++ projects/api/dxt-lossless-transform-bc1-api/src/c_api/transform/mod.rs
@@ -1,11 +1,10 @@
 //! C API for BC1 transform and untransform operations.
 
 pub mod builder;
-
-pub use builder::*;
+pub mod transform_context;
 
 use crate::c_api::error::{Dltbc1ErrorCode, Dltbc1Result};
-use crate::c_api::transform_context::{
+use crate::c_api::transform::transform_context::{
     Dltbc1TransformContext, get_detransform_details, get_transform_details,
 };
 use crate::transform::{transform_bc1_slice, untransform_bc1_slice};
